Add Authorization tests
