+++ src/test/java/org/apache/hadoop/hdfs/server/namenode/analytics/TestAuthorizationBase.java
@@ -25,11 +25,6 @@ import static org.hamcrest.core.StringContains.containsString;
 
 import java.io.IOException;
 import org.apache.commons.io.IOUtils;
-import org.apache.hadoop.hdfs.MiniDFSCluster;
-import org.apache.hadoop.hdfs.server.namenode.GSetGenerator;
-import org.apache.hadoop.hdfs.server.namenode.INode;
-import org.apache.hadoop.hdfs.server.namenode.INodeWithAdditionalFields;
-import org.apache.hadoop.util.GSet;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
@@ -37,28 +32,14 @@ import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.DefaultHttpClient;
 import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class TestAuthorization {
+public abstract class TestAuthorizationBase {
 
-  private static HttpHost hostPort;
-  private static HttpClient client;
-  private static WebServerMain nna;
+  protected static HttpHost hostPort;
+  protected static ApplicationMain nna;
 
-  @BeforeClass
-  public static void beforeClass() throws Exception {
-    GSetGenerator gSetGenerator = new GSetGenerator();
-    gSetGenerator.clear();
-    GSet<INode, INodeWithAdditionalFields> gset = gSetGenerator.getGSet((short) 3, 10, 500);
-    nna = new WebServerMain();
-    ApplicationConfiguration conf = new ApplicationConfiguration();
-    conf.set("ldap.enable", "false");
-    conf.set("authorization.enable", "true");
-    conf.set("nna.base.dir", MiniDFSCluster.getBaseDirectory());
-    nna.init(conf, gset);
-    hostPort = new HttpHost("localhost", 4567);
-  }
+  private HttpClient client;
 
   @AfterClass
   public static void tearDown() {
